# docs/swagger.yaml
definitions:
  response.CountrySummary:
    properties:
      country:
        type: string
      country_total_revenue:
        type: number
      products:
        items:
          $ref: '#/definitions/response.ProductRevenueSummary'
        type: array
    type: object
...
  response.ProductRevenueSummary:
    properties:
      product_name:
        type: string
      total_revenue:
        type: number
      transaction_count:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/revenue/country/summary:
    get:
      consumes:
      - application/json
      description: Returns a paginated list of countries with revenue & top products
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      - description: Number of top products per country
        in: query
        name: products
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.CountrySummary'
            type: array
      summary: Get country revenue summary
      tags:
      - revenue
...
swagger: "2.0"